[REF] website: factorize URL source into a service
When [1] introduced `html_editor`'s version of the URL autocomplete in
the link popup, it duplicated most of the existing code of the old
version, hoping that the old system would not need to be used anymore
in `html_builder`.
Unfortunately, the old system is about enabling the facility onto an
existing field rather than having an inline input inside the template.
It is therefore still being used.

This commit factorizes the URL source into a service in order to avoid
duplicating that logic.

[1]: https://github.com/odoo/odoo/pull/214005/commits/e1c8c86f1bb37a3c1bd231c113303ed94c402b99

task-4367641
