+++ addons/website/static/src/components/autocomplete_with_pages/url_autocomplete.js
@@ -1,10 +1,7 @@
 import { Component } from "@odoo/owl";
-import { rpc } from "@web/core/network/rpc";
-import { useChildRef } from "@web/core/utils/hooks";
+import { useChildRef, useService } from "@web/core/utils/hooks";
 import { AutoCompleteWithPages } from "@website/components/autocomplete_with_pages/autocomplete_with_pages";
 
-// TODO: we probably don't need it anymore after merging html_builder
-// see: https://github.com/odoo/odoo/pull/187091
 export class UrlAutoComplete extends Component {
     static props = {
         options: { type: Object },
@@ -15,6 +12,7 @@ export class UrlAutoComplete extends Component {
     static components = { AutoCompleteWithPages };
 
     setup() {
+        this.urlSource = useService("website_url_source");
         this.inputRef = useChildRef();
     }
 
@@ -37,49 +35,12 @@ export class UrlAutoComplete extends Component {
     get sources() {
         return [
             {
-                optionSlot: "option",
-                options: async (term) => {
-                    const makeItem = (item) => ({
-                        cssClass: "ui-autocomplete-item",
-                        label: item.label,
-                        onSelect: this.onSelect.bind(this, item.value),
-                    });
-
-                    if (term[0] === "#") {
-                        const anchors = await this.props.loadAnchors(
-                            term,
-                            this.props.options && this.props.options.body
-                        );
-                        return anchors.map((anchor) => makeItem({ label: anchor, value: anchor }));
-                    } else if (term.startsWith("http") || term.length === 0) {
-                        // avoid useless call to /website/get_suggested_links
-                        return [];
-                    }
-                    if (this.props.options.isDestroyed?.()) {
-                        return [];
-                    }
-                    const res = await rpc("/website/get_suggested_links", {
-                        needle: term,
-                        limit: 15,
-                    });
-                    const choices = [];
-                    for (const page of res.matching_pages) {
-                        choices.push(makeItem(page));
-                    }
-                    for (const other of res.others) {
-                        if (other.values.length) {
-                            choices.push({
-                                cssClass: "ui-autocomplete-category",
-                                data: { separator: true },
-                                label: other.title,
-                            });
-                            for (const page of other.values) {
-                                choices.push(makeItem(page));
-                            }
-                        }
-                    }
-                    return choices;
-                },
+                optionSlot: "urlOption",
+                options: this.urlSource.loadOptionsSource.bind({
+                    component: this,
+                    targetElement: this.props.options && this.props.options.body,
+                    onSelect: this.onSelect,
+                }),
             },
         ];
     }
