+++ addons/website/static/src/components/autocomplete_with_pages/url_autocomplete.xml
@@ -1,5 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
+    <t t-name="website.UrlAutoCompleteItem">
+        <div t-att-class="{
+            'fw-bold text-capitalize p-2': option.data.isCategory,
+        }">
+            <t t-if="option.data.icon">
+                <img t-att-src="option.data.icon" width="24px" height="24px" class="me-2 rounded"/>
+            </t>
+            <t t-out="option.label"/>
+        </div>
+    </t>
+
     <t t-name="website.UrlAutoComplete">
         <AutoCompleteWithPages
             input="inputRef"
@@ -9,15 +20,10 @@
             sources="sources"
             targetDropdown="props.targetDropdown"
         >
-            <t t-set-slot="option" t-slot-scope="optionScope">
-                <div t-att-class="{
-                    'fw-bold text-capitalize p-2': optionScope.data.separator,
-                }">
-                    <t t-if="optionScope.data.icon and optionScope.data.icon.length">
-                        <img t-att-src="optionScope.data.icon" width="24px" height="24px" class="me-2 rounded"/>
-                    </t>
-                    <t t-out="optionScope.label"/>
-                </div>
+            <t t-set-slot="urlOption" t-slot-scope="urlOptionScope">
+                <t t-call="website.UrlAutoCompleteItem">
+                    <t t-set="option" t-value="urlOptionScope"/>
+                </t>
             </t>
         </AutoCompleteWithPages>
     </t>
