+++ addons/website/static/src/xml/html_editor.xml
@@ -8,17 +8,6 @@
     </xpath>
 </t>
 
-<t t-name="website.AutoCompleteItem">
-    <div t-att-class="{
-        'fw-bold text-capitalize p-2': option.data.isCategory,
-    }">
-        <t t-if="option.data.icon">
-            <img t-att-src="option.data.icon" width="24px" height="24px" class="me-2 rounded"/>
-        </t>
-        <t t-out="option.label"/>
-    </div>
-</t>
-
 <t t-name="website.InputURLAutoComplete">
     <AutoCompleteInLinkPopover
         sources="sources"
@@ -32,7 +21,7 @@
         updateValue.bind="updateValue"
     >
         <t t-set-slot="urlOption" t-slot-scope="urlOptionScope">
-            <t t-call="website.AutoCompleteItem">
+            <t t-call="website.UrlAutoCompleteItem">
                 <t t-set="option" t-value="urlOptionScope"/>
             </t>
         </t>
